Better handle adding subsystems without data files
Created a enw class to handle all of the extra information. A subsystem
can now be labeled as such, but draw its data from files belong to a
different subsystem. This has been testing against processRuns(), as
well as for QA and partial merging.

In the process, QA was fully decoupled from spliiting out hists. This
was originally intended, but it needed a better implementedation.