Revise grid creation for better performance
Switch from a table to a floating div, which should be more responsive
than a table. In principle, this also allows for a dynamic selection of
how many images to display side by side. However, I opted to define the
number of columns during processing, and instead resize the images for a
particular display