Change over fully to ZODB classes
This completes basic database support. Closes #2. It will need to be
integrated as a server for full deployment, but this simple version
is fine for testing and a full version will be much easier to use
with a Docker image which is not yet completed.

Needed to adjust how the histGroupContainer stores hists from a
OrderedDict (which does not have a ZODB equivalent and is not as
nice as native ZODB) to a PersistentList(). This required some
adjustment in how the histGroup name is used, but it was resolved
because the dict key was almost always the selectedPattern name (the
exception was the Non-Subsystem hists).