Combine receiver scripts into initOVERWATCH.sh, and condense a single config file and a separate set of shared functions.
There is now a single init script initOVERWATCH.sh that (1) if run on a processing
machine: starts the receivers (if necessary) and runs processing, or (2) if run on a server
machine: starts the wsgi server. This includes automatic detection if the receivers
are running or if they need to be started, as well as a config option to force a restart
of all receivers. This init script is configured by a single config file configOVERWATCH.sh,
and utilizes a set of shared functions sharedFunctions.sh.