Update validation and bug fixes
Continues #18.

Also fixes bug accessing time slices directly. In particular, this can
generate a valid empty request for hist or hist groups, so we need to
process these into "None". However, an empty selection could also come
from a non-subsystem selection (ie "" in the group selection pattern).
To resolve this, we artificially assign the empty string non-subsystem
selection to a contrived string and filter it out later in the
validation.