Remove unneeded HLTMessageFormat switch
It is only needed if sending with the new AliZMQhelpers library and
receive with the old one. We never do this, so the switch is not
needed for our receiver. Thus, it is not set on advice from Mikolaj.
(This commit removes the setting, since I mistakenly set it
previously).