Symlink data instead of moving it
This hack allows rsyncing of the data with having to keep track of it or
moving it. It is necessary since the old version of overwatch has been
retired at CERN. Instead, the processing is now performed at Yale.

Note that this relates to the old version of overwatch, but it is
important to have available incase the system crashes.