Move executables into python package
Take advantage of setuptools to automatically generate the executables
to launch functions inside of a module. In each module, there is a
run.py which is responsible for running the module.

In the case of flask apps, the executable runs the flask development
server because the module can be run directly by uwsgi in production. In
particular, uwsgi just calls the module, not a function in it, while the
script calls a particular function. Thus, we can have both executables
to run when convenient and still be compatible with uwsgi.