Ensure additional hists are always processed
They would not be processed during the first pass of a run because the
histsAvailable dict was being assigned to histsInFile, which meant that
any entries in histsAvailable were lost.

In the process of implementing this, additional checks were added for
the possibility that a hist doesn't exist when a hist container attempts
to retrieve it. This can happen in the case of, for example, projections.
Even if the hist generally exists, the name may have changed from one
run to another (as is the case for some TPC hists)