Update configuration handling
Previously, we always loaded all configurations, regardless of the
module which was being executed, because the user config.yaml would
fail due to an anchor not always being available. This is resolved by
removing anchors from the user config.yaml that may be undefined (such
as those defined only in the webApp config) and loading the bcrypt log
rounds default directly in the config.py bcrypt yaml plugin. For most
users, this change doesn't really matter, but it cleans up the config
structure substantially.

This also allowed for integration tests for loading the configuration
to be implemented, which are also included here.