Update documentation system
Updated to the entire system as part of generally updating the
documentation. Some parts of the new system still need to be filled
out.

Unfortunately, sphinx and the related themes are fairly poorly
documented (such as the undocumented need for an `about.rst` to be able
to show a logo in the navigation). In general, rst is pretty painful,
but markdown based tools still don't seem to be mature enough to use
entirely.