Cleanup detectors `__init__.py` docs
Add additional information about the detectors specific functionality
and the fix for the import issue. For further information about the
import issue, `__init__.py` was accidentally deleted in
`overwatch.processing.detectors` when cleaning up out of date
documentation. This caused an import problem in python 2 (it
doesn't matter for py 3).