Cleanup obsolete TODO
It might be nice to include the file mode in the run information yaml
file, but it's also incredibly unlikely that we'll switch file modes,
so we can safely assume that it always be in cumulative mode. If it
changes, it can always be included in the run info yaml file.