Rename EMC options function for clarity
`generalEMCOptions` looks a lot like many of the plug-in function names,
but it isn't actually a plug-in function. Instead, it just applies to
all EMC histograms. So it was renamed to clarify that it is just a set
of general options. Furthermore, it requires the underlying hist and
canvas, so this is included for clarity (especially since it is
referenced in the more general documentation).

Previously, we had to worry about namespace collisions with a generic
names like `generalOptions`, but with the new function namespace, it is
just fine.