Clarify required arguments for web app validation
This basically duplicates the information from the web app, but it is
useful to provide it in both cases. It's useful in the web app since we
need to provide that information when calling the web app routing
function, and it's useful in the validation because that is where the
args are actually used.