Extract version in docs from overwatch.version
Use the same technique as in `setup.py` to read in the file and extract
the version number. I didn't directly import Overwatch because it could
cause the docs to fail (say, if ROOT is missing, since when determine
the version, we haven't yet put our shim into place). Admittedly, it may
be possible to fix this, but it is likely to cause trouble in the
future, so we will instead use the simple solution.