Modify Docker build process so it works for Travis
I had hoped that Travis would pickup and just use the intermediate layers
in the multi-stage build and not try to rebuild XRootD and ROOT, but
unfortunately this isn't the case. Consequently, we move back to the
base image building XRootD and ROOT, such that the standard Overwatch
doesn't need to build them. We these modifications, we should be able to
build the standard image on Travis.