Switch from `coveralls` to `codecov`
`coveralls` claims to successfully send the coverage, but for
whatever reason, the information never show up on their site.
So we'll use `codecov` instead.

For codecov to succeed, we have to run it in the docker container where
we run our tests. Otherwise, it won't resolve the paths successfully
and the coverage parsing will fail (without a clear error message).