Fix environment cleanup
Somehow, this cleanup was not sufficient such that the tests would
fail on python 2.7 when the entire deploy config was available in
an environment variable. The exact cause is not obvious because it
wouldn't fail if a specific test was run directly. However, doing a full
environment cleanup is sufficient to fix the issue.