Keep exec folders in the repo to help supervisor
Supervisor will fail if required directories do not exist. This can be
rather inconvenient, as we `overwatchDeploy` runs too late to create
them. To avoid this issue, we keep the directories in the repository, so
they'll always exist (assuming execution is started in the Overwatch
root directory, which is usually the case).