Improve receiver tests
Standardize the approach to loggingMixin (so it doesn't need to be
repeated called), and remove created the data folder from the DQM
receiver. No other overwatch executable does it, and it created the
folder whenever the module is imported (which meant it couldn't even
be mocked). So now it just won't do that