Improve `isRunOngoing()` logic
Now the run is considered ongoing if a file was received within the
last 5 minutes. This means that the ongoing state won't oscillate back
and forth depending on exactly when the last file arrived and was
processed. We do trade off lagging slightly behind the run state, but
given the propagation time from the HLT anyway, this seems quite
reasonable.