Ensure combined file updated for fileLocationSubsystem
While the combined file was updated for new files in a standard
subsystem, subsystems which were not their own file locations would be
left with out of date combined files. This now updates the combined
files (after the standard subsystems are updated, to ensure that any
combined file updates are propagated to those non file location
subsystems).

This should address OVERWATCH-PROCESSING-5 in Sentry