Remove combined file when converting subsystem
When moving from fileLocationSubsystem to normal subsystem, we forgot to
update the combined files. Now we remove the combined file when
converting.

We also ensure that fileLocationSubsystem combined files are updated
after the standard subsystem combined files, so they point to the most
recent ones.

This is another attempt to fix OVERWATCH-PROCESSING-7 on sentry.