Disable codecov patch check
Patch is supposed to measure the coverage of the particular commit or
PR. However, it consistently fails (because we don't have close to 100%
test coverage), and it doesn't really provide useful information at this
point. We can always just check the overall project coverage for
information on how it changed.