Clear subsystem histograms on conversion
When converting from a fileLocationSubsystem to a standard subsystem, we
cleared the files. However, that also means that the histograms can
change dramatically, and we didn't reset the subsystem. Now, we reset,
so the available histograms are reprocessed.

This should address OVERWATCH-PROCESSING-19 in sentry (and
also ensure that all of those are merged together in the future)