Quieter logs when histograms disappear from file to file
There appears to be some histograms which are not always sent in every
file. This happens sometimes for no clear reason. While we will continue
to log the information, we don't want to waste all of our sentry
notifications on this issue. So we change the log level to info