Fix docker push during Travis tag
It deployed to docker based on the branch name regardless of whether the
commit was tagged, but we didn't look for branch names in the form of
tags. So this modifies and documents the branches which should deploy
to docker, and the condition is satisfied for tags. Thus, now docker
images are deployed on tagging